changed volumePath to not use /ifs/volumes by default
Changed volume path to use /ifs/volumes if it's blank, but allow any other volume path to be specified. I opted to not add /ifs as a precursor because there is an ability within Isilon to create root symlinks so if something is heavily nested such as /ifs/containers/k8/app0/ could be directed to the root as /app0. Error checking is added to make sure critical directories cannot be  used as a root path.

Signed-off-by: Kendrick Coleman <kendrickcoleman@gmail.com>