Custom JSON Package
This patch introduces the ./api/json package, a clone of the StdLib
"encoding/json" package, with a few changes. Namely, the forked copy of
the JSON library introduces two new tags for struct fields named
"omitmarshal" and "omitunmarshal". These tags prevent a field from being
marshaled to or from JSON, respectively. This is a more granular
approach than the existing, blanket `json:"-"` solution.